refactor(S-3.01): extract handle_logout to auth/logout.rs
Move resolve_logout_target and handle_logout out of mod.rs.
resolve_logout_target re-exported under #[cfg(test)] for tests.
612 unit tests pass.
